Added tests to verify processing of escapeBlock escapeScheme where the escapeBlockEnd and escapeEscapeBlock characters are the same.
Added tests to verify processing of apostrophes (&apos;) within the document
and infoset data.

DAFFODIL-1923 DAFFODIL-1928
